% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getComposition}
\alias{getComposition}
\title{Get entries matching composition}
\usage{
getComposition(composition)
}
\value{
all composition matches
}
\description{
Get entries matching composition
}
\examples{
getComposition(hex=3&hexnac=5)
}
