% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getGlycanStructureGTs}
\alias{getGlycanStructureGTs}
\title{Get GTs involved in synthesis of structure}
\usage{
getGlycanStructureGTs(glytoucan_id)
}
\arguments{
\item{glytoucan_id}{}
}
\value{
lists of GTs
}
\description{
Get GTs involved in synthesis of structure
}
\examples{
getGlycanStructureGTs('G17689DH')

}
