% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getGlytoucanStructureDetails}
\alias{getGlytoucanStructureDetails}
\title{Get glycan structure details}
\usage{
getGlytoucanStructureDetails(glytoucan_id)
}
\value{
structure metadata
}
\description{
Get glycan structure details
}
\examples{
getGlytoucanStructureDetails('G17689DH')

}
