% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getProteoglycoforms}
\alias{getProteoglycoforms}
\title{Get Proteoglycoforms}
\usage{
getProteoglycoforms(uniprot)
}
\value{
all glycoforms for uniprot
}
\description{
Get Proteoglycoforms
}
\examples{
getProteoglycoforms('P14210-1')
}
