% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{searchComposition}
\alias{searchComposition}
\title{Get entries matching composition}
\usage{
searchComposition(composition)
}
\value{
all glycoproteins
}
\description{
Get entries matching composition
}
\examples{
searchComposition(hex=3&hexnac=5)
}
