% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{searchMass}
\alias{searchMass}
\title{Gets structures with a mass}
\usage{
searchMass(mass)
}
\arguments{
\item{mass}{Native mass}
}
\value{
glycan structure sequence details and mass
}
\description{
Gets structures with a mass
}
\examples{
searchMass(180)

}
