% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{searchMassGlycoprotein}
\alias{searchMassGlycoprotein}
\title{Gets structures and glycoproteins with a mass}
\usage{
searchMassGlycoprotein(mass)
}
\arguments{
\item{mass}{Native mass}
}
\value{
glycan structure sequences, mass and associated glycoproteins
}
\description{
Gets structures and glycoproteins with a mass
}
\examples{
searchMass(180)

}
