% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{searchMass}
\alias{searchMass}
\title{Gets structures with a mass}
\usage{
searchMass(mass)
}
\arguments{
\item{mass}{Native mass}
}
\value{
glytoucan
}
\description{
This function converts input temperatures in Fahrenheit to Celsius.
}
\examples{
searchMass(180)

}
